implements 'append' and 'reset' data capturing modes

+++ main.py
@@ -9,12 +9,14 @@ def run(**kwargs):
 
 def capture(**kwargs):
     path = kwargs["path"]
-    if os.path.exists(path):
-        shutil.rmtree(path)
+    mode = kwargs["mode"]
 
-    os.mkdir(path)
-    os.mkdir(os.path.join(path, "positives"))
-    os.mkdir(os.path.join(path, "negatives"))
+    if mode == "reset" and os.path.exists(path):
+        shutil.rmtree(path)
+    elif mode == "append" and not os.path.exists(path):
+        os.mkdir(path)
+        os.mkdir(os.path.join(path, "positives"))
+        os.mkdir(os.path.join(path, "negatives"))
 
     rec = Recognizer(src=0, haar_path="./haarcascade_frontalface.xml")
     rec.run()
@@ -31,15 +33,17 @@ commands = {"run": run, "capture": capture, "train": train, "serve": serve}
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="DeepFix human-robot interface")
     parser.add_argument("--command", dest="command", help="Name of the command to run", required=True)
+    parser.add_argument("--mode", dest="mode", default="append", help="Determines data capturing mode")
     parser.add_argument("--path", dest="path", default="images", help="Path to folder for captured images")
     args = parser.parse_args()
     
     cmd = args.command
-    pth = args.path
+    path = args.path
+    mode = args.mode
 
     if not cmd in allowed_cmds:
         print("{cmd}: command not found. --command must be one of {cmds}".format(cmd=cmd, cmds=allowed_cmds))
         sys.exit(-1)
 
-    fn = commands[args.command]
-    fn(path=pth)
+    action = commands[args.command]
+    action(path=path, mode=mode)
